extracted components and types to individual files
